Allow Stock Offload to handle fractional van stock
Piece-selling left a gap: a rep who sells a few pieces out of a carton
ends up holding a fractional amount of van stock (e.g. 1.75 cartons),
but Stock Offload only accepted whole-number quantities — that rep
could never return the exact amount they held, leaving a permanent
0.75-carton remainder stuck on their liability report with no way to
clear it.

Widened StockOffloadItem.quantity_declared/quantity_received to Float
(with the same Postgres ALTER COLUMN TYPE migration approach used for
the other piece-selling columns), and updated offload_submit()/
offload_confirm() to parse floats instead of ints. The offload form's
quantity step now matches the product's pieces_per_unit granularity so
the input doesn't reject a valid fractional value.

Verified via an isolated DB-copy test: submitting and confirming a
1.75-carton offload moves the exact amount, and a warehouse manager
trying to over-receive still gets clamped to what was actually declared.

+++ models/van_management.py
@@ -102,7 +102,10 @@ class StockOffloadItem(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     offload_id = db.Column(db.Integer, db.ForeignKey('stock_offloads.id'), nullable=False)
     product_id = db.Column(db.Integer, db.ForeignKey('products.id'), nullable=False)
-    quantity_declared = db.Column(db.Integer, nullable=False, default=0)
-    quantity_received = db.Column(db.Integer, nullable=True)
+    # Float, not Integer — a rep can now hold a fractional amount of van
+    # stock after selling by the piece, and must be able to offload exactly
+    # that amount back to the warehouse rather than being stuck rounding down.
+    quantity_declared = db.Column(db.Float, nullable=False, default=0)
+    quantity_received = db.Column(db.Float, nullable=True)
 
     product = db.relationship('Product', foreign_keys=[product_id])
